+++ src/controllers/tweet.controller.js
@@ -3,7 +3,7 @@ import { Tweet } from "../models/tweet.model.js"
 import { User } from "../models/user.model.js"
 import { ApiError } from "../utils/ApiError.js"
 import { ApiResponse } from "../utils/ApiResponse.js"
-import { asyncHandler } from "../utils/asyncHandler.js"
+import asyncHandler from "../utils/asyncHandler.js"
 
 const createTweet = asyncHandler(async (req, res) => {
     const { content } = req.body
@@ -44,6 +44,10 @@ const updateTweet = asyncHandler(async (req, res) => {
         { new: true }
     )
 
+    if (!tweet) {
+        throw new ApiError(404, "Tweet not found")
+    }
+
     return res
         .status(200)
         .json(
@@ -53,6 +57,12 @@ const updateTweet = asyncHandler(async (req, res) => {
 
 const deleteTweet = asyncHandler(async (req, res) => {
     const { tweetId } = req.params
+
+    const tweet = await Tweet.findById(tweetId)
+
+    if (!tweet) {
+        throw new ApiError(404, "Tweet not found")
+    }
 
     await Tweet.findByIdAndDelete(tweetId)
